Skip if it is a specific value

+++ bin/update_phpbb.sh
@@ -8,12 +8,12 @@ HERE=$(dirname "$0")
 cd "$HERE/../area51-phpbb3/phpBB"
 git fetch "$REMOTE"
 
-if [ `git rev-parse "$BRANCH"` != `git rev-parse "$REMOTE/$BRANCH"` ]
+if [ `git rev-parse "$BRANCH"` != `git rev-parse "$REMOTE/$BRANCH"` ] && [ `git rev-parse "$BRANCH"` != beee215bc2d28159b37d5e24e37598fd3d03edbd ]
 then
 	git reset --hard "$REMOTE/$BRANCH"
 	../composer.phar install --no-dev --optimize-autoloader
 	bin/phpbbcli.php --safe-mode db:migrate
 	bin/phpbbcli.php --safe-mode cache:purge
-	#rm cache/production/*.{lock,php}
-	#rm -r cache/production/twig/
+	rm cache/production/*.{lock,php}
+	rm -r cache/production/twig/
 fi
